requirements.txt
selenium 
pytest123 
pytest-html 